+++ xarray/datatree_/datatree/mapping.py
@@ -189,14 +189,15 @@ def map_over_subtree(func: Callable) -> Callable:
             *args_as_tree_length_iterables,
             *list(kwargs_as_tree_length_iterables.values()),
         ):
-            node_args_as_datasetviews = [
-                a.ds if isinstance(a, DataTree) else a for a in all_node_args[:n_args]
+            node_args_as_datasets = [
+                a.to_dataset() if isinstance(a, DataTree) else a
+                for a in all_node_args[:n_args]
             ]
-            node_kwargs_as_datasetviews = dict(
+            node_kwargs_as_datasets = dict(
                 zip(
                     [k for k in kwargs_as_tree_length_iterables.keys()],
                     [
-                        v.ds if isinstance(v, DataTree) else v
+                        v.to_dataset() if isinstance(v, DataTree) else v
                         for v in all_node_args[n_args:]
                     ],
                 )
@@ -204,7 +205,7 @@ def map_over_subtree(func: Callable) -> Callable:
 
             # Now we can call func on the data in this particular set of corresponding nodes
             results = (
-                func(*node_args_as_datasetviews, **node_kwargs_as_datasetviews)
+                func(*node_args_as_datasets, **node_kwargs_as_datasets)
                 if not node_of_first_tree.is_empty
                 else None
             )